Do everything as float

+++ fahbench/cmd/CommandLineUpdater.cpp
@@ -1,12 +1,12 @@
 #include <iostream>
 #include "CommandLineUpdater.h"
 
-void CommandLineUpdater::progress(int i, int numSteps, double ns_per_day) {
+void CommandLineUpdater::progress(int i, int numSteps, float ns_per_day) {
     if (i == numSteps) {
         std::cout << std::string(80, ' ') << std::endl;
         return;
     }
-    double percent = (100.0 * i) / numSteps;
+    float percent = (100.0 * i) / numSteps;
     std::cout << boost::format("Progress: %1$5.1f%%  estimate: %2$6.2f ns/day") % percent % ns_per_day << "\r" << std::flush;
 }
